Add support for reCAPTCHA verification when using the perspectiveapi-simple-server (#92)
This demo issues requests against an instance of the [perspectiveapi-simple-server](https://github.com/conversationai/perspectiveapi-simple-server) to fetch data from the Perspective API. Recently, we introduced optional reCAPTCHA verification to the simple-server.

With this change, we optionally inject the `RecaptchaV3Service` from the [ng-recaptcha](https://github.com/conversationai/perspectiveapi-simple-server) package to fetch and append a reCAPTCHA token to requests to the simple-server. Note that since this is an optional dependency, reCAPTCHA verification is not enabled by default. To enable the verification flow, developers must provide a `recaptchaConfig` in `server_config.json` and add `RecaptchaV3Service` and `RECAPTCHA_V3_SITE_KEY` to the list of providers for the module importing this module.